backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.0
openai>=1.3.0
pytest>=7.4.0
httpx>=0.25.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
fastapi-users[sqlalchemy]>=12.1.0
slowapi>=0.1.9
redis>=5.0.0 